added ForbiddenException upon remov

+++ src/apatrments/apatrments.service.ts
@@ -1,14 +1,21 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import {
+  ForbiddenException,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
 import { InjectModel } from '@nestjs/mongoose';
 import { CreateApatrmentDto } from './dto/create-apatrment.dto';
 import { UpdateApatrmentDto } from './dto/update-apatrment.dto';
 import { Apartment } from './schemas/apatrment.schema';
 import { Model } from 'mongoose';
+import { ContactsService } from 'src/contacts/contacts.service';
+import { Contact } from 'src/contacts/schemas/contact.schema';
 
 @Injectable()
 export class ApatrmentsService {
   constructor(
     @InjectModel(Apartment.name) private apartmentModel: Model<Apartment>,
+    private contactsService: ContactsService,
   ) {}
   async create(createApatrmentDto: CreateApatrmentDto) {
     const createdApartment = new this.apartmentModel(createApatrmentDto);
@@ -35,9 +42,23 @@ export class ApatrmentsService {
   }
 
   async remove(id: string) {
+    const residentInTheApartment: Contact | null =
+      await this.contactsService.findOneByApartment(id);
+    if (residentInTheApartment) {
+      const { City, Street, BuildingNumber, AppartmentNumber } =
+        residentInTheApartment.Apartment;
+      const { FirstName, LastName } = residentInTheApartment;
+
+      throw new ForbiddenException(
+        `cannot delete apartment: ${Street} ${BuildingNumber} apt. ${AppartmentNumber}, ${City} which habitad by resident ${FirstName} ${LastName}`,
+      );
+    }
+
     const removedApartment = await this.apartmentModel.findByIdAndRemove(id);
-    if (!removedApartment)
-      throw new NotFoundException(id, 'apartment not found');
     return removedApartment;
+  }
+
+  async removeAll() {
+    return await this.apartmentModel.remove();
   }
 }
